requirements.txt
fastapi==0.95.2
uvicorn
pillow==9.5.0
ultralytics==8.0.145
python-multipart==0.0.6
streamlit==1.25.0
numpy==1.24.3
